getting frmnlist in dashboard

+++ BIS.Common/Entities/Corps.cs
@@ -6,14 +6,14 @@ using System.Threading.Tasks;
 
 namespace BIS.Common.Entities
 {
-    public class Corps : CommonModel
-    {
-        public string Name { get; set; }
+	public class Corps : CommonModel
+	{
+		public string Name { get; set; }
 
-    }
-    public class Divisons : CommonModel
-    {
-        public int CorpsId { get; set; }
-        public string Name { get; set; }
-    }
+	}
+	public class Divisons : CommonModel
+	{
+		public long CorpsId { get; set; }
+		public string Name { get; set; }
+	}
 }
